Added viscous torsion hydro load, different flow types, and fixed weight and buoyancy forces
Added a viscous load to the body that is proportional to the square of the rotation in the body-z direction. Also fixed the weight and buoyancy implementation so that you get a hydrostatic moment when the two forces are not collocated. Changed the plots to RPM for the body rotation so it is easier to compare to the rotor rotation. Added a few different flow types to the fluid class (still all uniform). Put the tether load computation on the tether class. Added a test for the moment coefficient for the viscous torque on the body.