Exposed some internal parameters
Exposed the normal and axial coefficents for the viscous damping on the body. Created a modification parameter for the viscous torsion. Added a parameter to set the viscous damping coefficient between the rotors and body. Also added a parameter that lets you control whether or not to use the higher-fidelity computation of viscous torsion coefficent.