Added points in the S814 force coefficient curve. Added pitch blade property.  Changed force coefficient argument to AoA. Fully implemented Ning BEMT algorithm.
As the amin comment says. Also added a template input file. Added a tangential flow factor in addition to the existing axial flow factor. Improved comments here and there. Added a test script for the BEMT algorithm. Moved the output of the tether analysis to the right place. Added a script to make some torque curves. Also added a torque curve visualization method to the rotor class. Committing sandbox because why not, but that's mostly for me to do algebra that I'm too lazy to do by hand.