Bug fix and removed nonfunctional scripts from the main folder
Fixed the tangential flow factor bug in the setup method of the simulation class. added a couple of demo input files. Removed the old script m files from the old single rotor days. At this point, the coaxial dual rotors assumption is pretty well embedded in the code. Generalizing to abitrary number of rotors and orientations in the body frame will require effort.