Ballast integration and middle-road assumptions for EOMs
Added a "ballast" mass that can be placed anywhere in the body frame to simulate a ballast. Correctly modifies the mass, center of mass, anc inertia matrix. Also relaxed the assumption that the body frame is along the principal axes. In doing that, I realized that the equations in the state file had more generality than necessary. The equations now assume that the i-k plane is a plane of symmetry and the body k axis is coincident with the axis of the slender vehilce body. Also added a catch for if no ballast is specified.