Bug fixes
Foxed bug in rotor class that produced incorrect BEMT plots for the power and torque coefficients. Cosmetic change in tether class. Vehicle erroneously had the body forces commented out and the force/length was not getting multiplied by length to get force. Adding script to make the input file specifically for the tactical scale just so I have a record of it. Also adding a qick renaming script that can be use to rename sim files for whatever reason.