Changes for hydrostatic analyses
Added capability to modify cases in the main execution loop, including a method to automatically computer hydrostatic ICs in simulation class. Added a figure to the analyze steady state script and made changes to titles and labels. Overhaul of the input file maker script to make it easier to keep records of what what is in the input files.