Functionality improvements
Added degree offset to the main script. Added a title control parameter to makePlots method and support for call the static method from a sub folder. Adding ballast location property to the vehicle, but it will be empty in existing .mat files until we populate it (new sims will have it). Adding inter-sweep visulation tools that will likely get consolodated later. AnalyzeSteadyState makes better plots now. Commented-out code in the extendSimulations script is for checking for the second equilibrium when the lee rotor starts spinning the wrong way. Adding a function to support the inter-sweep analyses.