Preparation for multi-link tether activation
Clean-up in DualRotorSim, more parameters in the makeMovies method in simulation. Changed plotting in the analyzeAcrossFlowSpeeds script to get better figures for a manuscript. Same with analyzeSteadyState. Cleanup in extendSimulations script. StaticOrientation is a cluster of hack. This is the script that I wrote to compare sims to theory in static flow for a manuscript. It's not an important script, just here for record.